+++ EmdFlat/flow_t.cs
@@ -4,8 +4,8 @@ namespace EmdFlat
 {
     public struct flow_t
     {
-        public int from;      /* Feature number in signature 1 */
-        public int to;        /* Feature number in signature 2 */
-        public float amount;  /* Amount of flow from "from" to "to" */
+        public int from;       /* Feature number in signature 1 */
+        public int to;         /* Feature number in signature 2 */
+        public double amount;  /* Amount of flow from "from" to "to" */
     }
 }
